+++ examples/dummy-service/server.sh
@@ -1,6 +1,3 @@
 #!/bin/bash
 echo "Starting dummy server on port $PORT"
-while true; do
-  echo "Running..."
-  sleep 5
-done
+python3 -m http.server $PORT